Include prefix method unified in TaskUtils class in nextflow.util
