+++ modules/nextflow/src/main/groovy/nextflow/util/TaskUtils.groovy
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2013-2024, Seqera Labs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package nextflow.util
+
+import groovy.transform.CompileStatic
+
+/**
+ * Common task utilities.
+ *
+ * @author Jorge Ejarque <jorge.ejarque@seqera.io>
+ */
+@CompileStatic
+class TaskUtils {
+
+    static String includeTowerPrefix(String name, Map<String,String> environment) {
+        if ( environment?.containsKey('TOWER_WORKFLOW_ID') )
+            return "tw-"+environment['TOWER_WORKFLOW_ID'] + "-" + name
+        else
+            return name
+    }
+}
