+++ plugins/nf-azure/src/test/nextflow/cloud/azure/batch/AzBatchServiceTest.groovy
@@ -739,17 +739,4 @@ class AzBatchServiceTest extends Specification {
         [managedIdentity: [clientId: 'client-123']]     | 'client-123'
     }
 
-    def 'should include the tower prefix'() {
-        given:
-        def name = 'job_1'
-
-        expect:
-        AzBatchService.includeTowerPrefix(name, ENV) == EXPECTED
-
-        where:
-        ENV                         | EXPECTED
-        [TOWER_WORKFLOW_ID: '1234'] | "tw-1234-job_1"
-        [:]                         | "job_1"
-    }
-
 }
